# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CandidatePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateSummaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CandidatePage",
        "type": "object"
      },
...
      "DocumentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DocumentList",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Stable, for your code",
                "type": "string"
              },
              "details": {
                "description": "Depends on code",
                "type": "object"
              },
              "message": {
                "description": "For people; may change",
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "EvaluationList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EvaluationList",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A sign-in token from the company identity provider, or from /dev/token in dev.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Talent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/candidates": {
      "get": {
        "description": "Candidates in your scope, newest first, as summaries.\n\n`source` narrows by where a record came from: `?source=cv_upload&source=public_apply` is the\npeople who have arrived since the migration, without the 5,140 that came from the sheet.",
        "operationId": "list_candidates_v1_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "requisition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requisition Id"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "description": "Where the record came from; repeat for more than one",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "tai_master",
                      "cv_upload",
                      "manual_entry",
                      "public_apply"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Where the record came from; repeat for more than one",
              "title": "Source"
            }
          },
          {
            "description": "true: only archived records; false: only the ones still in play",
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true: only archived records; false: only the ones still in play",
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Candidates",
        "tags": [
          "candidates"
        ]
      },
      "post": {
        "description": "A candidate you found, typed in by hand. Every field is recorded as manual and unchecked,\nwith your name, and the candidate joins the review queue until a person checks them. Fields\nleft out or empty are not recorded. The response carries ids only.",
        "operationId": "enter_candidate_v1_candidates_post",
        "parameters": [
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnteredCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enter Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
    "/v1/candidates/search": {
      "post": {
        "description": "Finds candidates in your scope by full name, email or phone, sent in the body and never in\nthe URL. Every value sent must match. Matching is exact once case, spacing and phone formats are\ntidied; it is not a fuzzy or duplicate search. Returns summaries, newest first.",
        "operationId": "search_candidates_v1_candidates_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateSearchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Candidates",
        "tags": [
          "candidates"
        ]
      }
    },
    "/v1/candidates/{candidate_id}": {
      "get": {
        "description": "One candidate, every field with its source and verification (BR-201, BR-202).",
        "operationId": "get_candidate_v1_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/documents": {
      "get": {
        "description": "The candidate's CVs, oldest first, with how reading each one ended. Every listing is\nrecorded with your name.",
        "operationId": "list_documents_v1_candidates__candidate_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/evaluations": {
      "get": {
        "description": "Every evaluation of a candidate, newest first (BR-303).",
        "operationId": "list_evaluations_v1_candidates__candidate_id__evaluations_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Evaluations",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/fields/{field}/verification": {
      "post": {
        "description": "You checked one field. Records a new, verified row with your name and the time; the old\nrow stays. Send `value` when you corrected it. Once no field is left unchecked, the\ncandidate's unverified_candidate review item is resolved as checked, by you.",
        "operationId": "verify_field_v1_candidates__candidate_id__fields__field__verification_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "field",
            "required": true,
            "schema": {
              "enum": [
                "full_name",
                "phone",
                "whatsapp",
                "email",
                "location",
                "current_title",
                "current_employer",
                "education",
                "graduation_year",
                "years_experience",
                "age",
                "profile_url"
              ],
              "title": "Field",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldCheckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Field",
        "tags": [
          "candidates"
        ]
      }
    },
...
    "/v1/documents/{document_id}/file": {
      "get": {
        "description": "The original file, exactly as it was uploaded. Every download is recorded with your name.",
        "operationId": "download_document_v1_documents__document_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file, byte for byte as uploaded"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Document",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}